Added Second Button and First Tutorial Page
The newly-created button gives the user to go to the tutorial. The existing button has been changed into a button where, on click, sends the user directly to authenticate via log on or registration. This allows the user to choose wether they want to follow the tutorial or simply log on/register to the application.